Addressing C++ Prefast issues (#92)
* Addressing C++ Prefast issues

This PR addresses prefast issues:
- C6387
- C26451
- C28251
- C6011
- C6385

+++ CppCustomVisualizer2/dll/ChildVisualizer.h
@@ -79,7 +79,7 @@ private:
         _In_ DkmString* pEvalText,
         _In_ DkmString* pDisplayName,
         _In_ DkmString* pType,
-        _In_ unsigned long long index,
+        _In_ UINT32 index,
         _Deref_out_ DkmChildVisualizedExpression** ppResult
     );
 };
